# api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "V.SMART API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/currencies": {
      "get": {
        "tags": [
          "Currency"
        ],
        "summary": "M2-B02 — the reference implementation of the paged list contract (ADR-002 §2 and its\r\nM2-B02 addendum). Every future list endpoint copies this shape: one\r\n`[FromQuery]` typed query record in, one V.SMART.Api.Contracts.PagedResult`1 out, the\r\nservice's filter dictionary produced by V.SMART.Api.Contracts.FilterDictionaryAdapter and\r\nnever seen on the wire.\r\n            \r\n\r\nThe controller-local paged-response record it replaces is deleted: 60–80copies of it would become 60–80 interfaces in the generated TypeScript client. TheJSON is unchanged — the four property names were already ADR-002's.\r\n<b>Behaviour change to note:</b> the default `pageSize` is now thecontract-wide 20, where this endpoint alone previously defaulted to 10. Callers thatsend `pageSize` explicitly are unaffected.\r\nValidation is declarative on V.SMART.Api.Contracts.CurrencyQuery, so an invalid`pageNumber`/`pageSize`/`sort`/date range is rejected by M2-A06's`InvalidModelStateResponseFactory` before this method runs. The explicit`ModelState` guard below is the belt-and-braces path for a caller that invokes theaction directly (unit tests), and matches `Create`/`Update`.",
        "operationId": "getCurrencies",
        "parameters": [
          {
            "name": "currName",
            "in": "query",
            "description": "Case-insensitive contains match on the currency name. Wire name `currName`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdBy",
            "in": "query",
            "description": "Case-insensitive contains match on the creating user. Wire name `createdBy`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "description": "Inclusive lower bound on `CreatedDate`, applied at the start of the day.\r\n            Wire name `fromDate`.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "description": "Inclusive upper bound on `CreatedDate`, applied at the end of the day.\r\n            Wire name `toDate`.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "1-based page index. Default 1. Wire name `pageNumber`.",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Rows per page. Default 20, maximum 100. Wire name `pageSize`.",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated sort fields, `-` prefix for descending — e.g.\r\n`-createdDate,currName`. Omit it to keep the resource's existing default ordering.\r\nValidated against V.SMART.Api.Contracts.PagedQuery.SortableFields; an unknown field is a 400 that lists\r\nthe permitted values. Wire name `sort`.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVMPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVMPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVMPagedResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Currency"
        ],
        "summary": "Creates a currency. A duplicate name is the service's refusal, surfaced as 409 with the\r\nservice's message verbatim in `title` (ADR-002 §4); a malformed body is 400 with the\r\nper-field `errors` dictionary.",
        "operationId": "createCurrency",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurrencyVM"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CurrencyVM"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CurrencyVM"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVM"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVM"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVM"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/currencies/{id}": {
      "get": {
        "tags": [
          "Currency"
        ],
        "summary": "One currency by id. Answers 404 in the canonical problem shape when the row does not\r\nexist.",
        "operationId": "getCurrencyById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVM"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVM"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVM"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Currency"
        ],
        "summary": "Updates the currency identified by the route id. The route id is what the service acts\r\non; any `currId` in the body is ignored.",
        "operationId": "updateCurrency",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurrencyVM"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CurrencyVM"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CurrencyVM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVM"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVM"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVM"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Currency"
        ],
        "summary": "Deletes a currency. A delete guard's refusal is 409 carrying the service's message\r\nverbatim (BR-SO-001), not 400.",
        "operationId": "deleteCurrency",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CurrencyVM": {
        "required": [
          "currName",
          "currSub",
          "symbol"
        ],
        "type": "object",
        "properties": {
          "currId": {
            "type": "integer",
            "format": "int32"
          },
          "currName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "currSub": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "symbol": {
            "minLength": 1,
            "pattern": "[$€₹¥£₩₿₽]",
            "type": "string"
          },
          "isSystemDefined": {
            "type": "boolean"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CurrencyVMPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurrencyVM"
            },
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "The single paged-list response body for every list endpoint in this API (M2-B02).\r\n            \r\n\r\nThe property names are ADR-002 §2's `{ items, totalCount, pageNumber, pageSize }`verbatim; with ASP.NET Core's default camel-case JSON policy they serialise to exactlythose four names. This is <b>one generic type</b>, deliberately replacing thecontroller-local paged-response record pattern (the one `CurrencyController` used to declare): 60–80 structurally identicalresponse records would become 60–80 distinct interfaces in the generated TypeScriptclient (M2-B10).\r\n<b>Why there is no `totalPages`.</b> ADR-002 §2 names four properties and thegenerated client is frozen at M2-B03, so adding a fifth here is a contract decision, not aconvenience. It is derivable on the client from `totalCount` and `pageSize`.Note that an unrelated, unused `V.SMART.Shared.ViewModels.PagedResult<T>`(`RejectionMasterVM.cs:33-40`) does carry `TotalPages` — it has zero referencesand is not this type. Do not confuse the two; if a file needs both namespaces it mustalias one.\r\n`TotalCount` is the count of the <b>filtered, unpaged</b> query — see`CurrencyService.cs:76`, where `CountAsync()` runs after the filters and before`Skip`/`Take`. Every service in `BusinessLayer/` follows that convention(INV-002); a list endpoint that reports anything else is a defect."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}